requirements.txt
# requirements.txt
Django~=5.2
django-extensions
gunicorn
pandas
scikit-learn
matplotlib
numpy
kafka-python
confluent-kafka
dataclasses-json
psycopg2-binary
dj-database-url
